refactor light-update service
